+++ src/include/cxxgui/text.hpp
@@ -7,7 +7,10 @@ namespace cxxgui {
 
         std::string t = "";
         font_t* internal_font = nullptr;
-        GLuint texture_id = 0;
+        GLuint* texture_id = nullptr;
+        int* widths = nullptr;
+        int* heights = nullptr;
+        size_t texture_count = 0;
         float width = 0;
         float height = 0;
 
@@ -21,7 +24,14 @@ namespace cxxgui {
         text(std::string str) : t(str) { internal_font = cxxgui::body; }
         text(const char* str, font_t* f) : t(str), internal_font(f) { }
         text(std::string str, font_t* f) : t(str), internal_font(f) {}
-        ~text() { if(texture_id != 0) glDeleteTextures(1, &texture_id); }
+        ~text() {
+            if(texture_id != nullptr) {
+                glDeleteTextures((int)texture_count, texture_id);
+                delete[] texture_id;
+                delete[] widths;
+                delete[] heights;
+            }
+        }
 
         /*
          * Update text